MAINT: Splitting up plans.py into a directory with separate plans.
