+++ hxrsnd/__init__.py
@@ -1,5 +1,5 @@
 import logging
-from .plans import rocking_curve, maximize_lorentz
+from .plans.alignment import rocking_curve, maximize_lorentz
 
 from ._version import get_versions
 __version__ = get_versions()['version']
