+++ hxrsnd/tests/__init__.py
@@ -1,2 +1,2 @@
 from .test_states import Puck
-from .test_plans  import crystal, Diode
+from .conftest  import crystal, Diode
